+++ scripts/testing/debug_draw.gd
@@ -4,8 +4,7 @@ var rects = []
 
 func _process(_delta):
 	queue_redraw()
-
-
+	
 func _draw():
 	for r in rects:
 		draw_rect(r[0], r[1])